fixed large waning compartments not spanning all strains
if waning compartments were large enough an issue would arise where historical strain breakpoints could happen in the middle of the compartment. For example with a compartment of 100 days starting at the beginning of 2022, the first month or so would be inside of omicron timeframe, while the rest in delta. This needed to be reflected in the code.